requirements.txt
fastapi==0.115.0
uvicorn==0.31.1
jinja2==3.1.4
python-multipart==0.0.12
fastapi-cors==0.0.6
websockets==13.0
python-telegram-bot==21.8
aiogram==3.15.0
telethon==1.39.0
urllib3==2.4.0
requests==2.32.3
psycopg2-binary==2.9.10
init_data_py==0.2.6
asyncpg==0.30.0